+++ x/authority/types/message_update_chain_info_test.go
@@ -32,12 +32,12 @@ func TestMsgUpdateChainInfo_ValidateBasic(t *testing.T) {
 				Chains: []chains.Chain{
 					{
 						ChainId:     0,
-						ChainName:   chains.ChainName_empty,
 						Network:     chains.Network_optimism,
 						NetworkType: chains.NetworkType_testnet,
 						Vm:          chains.Vm_evm,
 						Consensus:   chains.Consensus_op_stack,
 						IsExternal:  true,
+						Name:        "foo",
 					},
 				},
 			}),