+++ zetaclient/chains/bitcoin/observer/inbound.go
@@ -233,13 +233,18 @@ func (ob *Observer) ProcessInboundTrackers(ctx context.Context) error {
 
 	for _, tracker := range trackers {
 		ob.logger.Inbound.Info().
-			Msgf("checking tracker with hash :%s and coin-type :%s ", tracker.TxHash, tracker.CoinType)
+			Str("tracker.hash", tracker.TxHash).
+			Str("tracker.coin-type", tracker.CoinType.String()).
+			Msgf("checking tracker")
 		ballotIdentifier, err := ob.CheckReceiptForBtcTxHash(ctx, tracker.TxHash, true)
 		if err != nil {
 			return err
 		}
 		ob.logger.Inbound.Info().
-			Msgf("Vote submitted for inbound Tracker, Chain : %s,Ballot Identifier : %s, coin-type %s", ob.Chain().ChainName, ballotIdentifier, coin.CoinType_Gas.String())
+			Str("inbound.chain", ob.Chain().Name).
+			Str("inbound.ballot", ballotIdentifier).
+			Str("inbound.coin-type", coin.CoinType_Gas.String()).
+			Msgf("Vote submitted for inbound Tracker")
 	}
 
 	return nil
@@ -470,5 +475,41 @@ func GetBtcEvent(
 			TxHash:      tx.Txid,
 		}, nil
 	}
+	return nil, nil
+}
+
+// GetBtcEventWithWitness either returns a valid BTCInboundEvent or nil.
+// This method supports data with more than 80 bytes by scanning the witness for possible presence of a tapscript.
+// It will first prioritize OP_RETURN over tapscript.
+func GetBtcEventWithWitness(
+	client interfaces.BTCRPCClient,
+	tx btcjson.TxRawResult,
+	tssAddress string,
+	blockNumber uint64,
+	logger zerolog.Logger,
+	netParams *chaincfg.Params,
+	depositorFee float64,
+) (*BTCInboundEvent, error) {
+	// first check for OP_RETURN data
+	event, err := GetBtcEvent(
+		client,
+		tx,
+		tssAddress,
+		blockNumber,
+		logger,
+		netParams,
+		depositorFee,
+	)
+
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to get btc event")
+	}
+
+	if event != nil {
+		return event, nil
+	}
+
+	// TODO: integrate parsing script
+
 	return nil, nil
 }