+++ zetaclient/chains/bitcoin/signer/signer.go
@@ -433,7 +433,7 @@ func (signer *Signer) TryProcessOutbound(
 		return
 	}
 	logger.Info().
-		Msgf("Key-sign success: %d => %s, nonce %d", cctx.InboundParams.SenderChainId, chain.ChainName, outboundTssNonce)
+		Msgf("Key-sign success: %d => %s, nonce %d", cctx.InboundParams.SenderChainId, chain.Name, outboundTssNonce)
 
 	// FIXME: add prometheus metrics
 	_, err = zetacoreClient.GetObserverList(ctx)
@@ -445,7 +445,7 @@ func (signer *Signer) TryProcessOutbound(
 	if tx != nil {
 		outboundHash := tx.TxHash().String()
 		logger.Info().
-			Msgf("on chain %s nonce %d, outboundHash %s signer %s", chain.ChainName, outboundTssNonce, outboundHash, signerAddress)
+			Msgf("on chain %s nonce %d, outboundHash %s signer %s", chain.Name, outboundTssNonce, outboundHash, signerAddress)
 
 		// try broacasting tx with increasing backoff (1s, 2s, 4s, 8s, 16s) in case of RPC error
 		backOff := broadcastBackoff
@@ -455,7 +455,7 @@ func (signer *Signer) TryProcessOutbound(
 			if err != nil {
 				logger.Warn().
 					Err(err).
-					Msgf("broadcasting tx %s to chain %s: nonce %d, retry %d", outboundHash, chain.ChainName, outboundTssNonce, i)
+					Msgf("broadcasting tx %s to chain %s: nonce %d, retry %d", outboundHash, chain.Name, outboundTssNonce, i)
 				backOff *= 2
 				continue
 			}
@@ -472,7 +472,7 @@ func (signer *Signer) TryProcessOutbound(
 			)
 			if err != nil {
 				logger.Err(err).
-					Msgf("Unable to add to tracker on zetacore: nonce %d chain %s outboundHash %s", outboundTssNonce, chain.ChainName, outboundHash)
+					Msgf("Unable to add to tracker on zetacore: nonce %d chain %s outboundHash %s", outboundTssNonce, chain.Name, outboundHash)
 			}
 			logger.Info().Msgf("Broadcast to core successful %s", zetaHash)